Initial update

+++ View/frmCategoryView.cs
@@ -57,20 +57,30 @@ namespace IM_System.View
                 frm.txtName.Text = Convert.ToString(guna2DataGridView1.CurrentRow.Cells["dgvname"].Value);
  
                 MainClass.BlurBackground(frm);
+                LoadData();// Refresh data after update
             }
             //Delete
             if (guna2DataGridView1.CurrentCell.OwningColumn.Name == "dgvDel")
             {
-                int id = Convert.ToInt32(guna2DataGridView1.CurrentRow.Cells["dgvid"].Value);
-                string qry = "Delete from Category where catID = " + id + "";
-                Hashtable ht = new Hashtable();
-                if (MainClass.SQl(qry, ht) > 0)
+                //Confirm Before delete
+                guna2MessageDialog1.Buttons = Guna.UI2.WinForms.MessageDialogButtons.YesNo;
+                guna2MessageDialog1.Icon = Guna.UI2.WinForms.MessageDialogIcon.Information;
+
+                if (guna2MessageDialog1.Show("Are you sure you want to delete") == DialogResult.Yes)
                 {
-                    guna2MessageDialog1.Buttons = Guna.UI2.WinForms.MessageDialogButtons.OK;
-                    guna2MessageDialog1.Icon = Guna.UI2.WinForms.MessageDialogIcon.Information;
-                    guna2MessageDialog1.Show("Delled successfuly..");
-                    LoadData();
+
+                    int id = Convert.ToInt32(guna2DataGridView1.CurrentRow.Cells["dgvid"].Value);
+                    string qry = "Delete from Category where catID = " + id + "";
+                    Hashtable ht = new Hashtable();
+                    if (MainClass.SQl(qry, ht) > 0)
+                    {
+                        guna2MessageDialog1.Buttons = Guna.UI2.WinForms.MessageDialogButtons.OK;
+                        guna2MessageDialog1.Icon = Guna.UI2.WinForms.MessageDialogIcon.Information;
+                        guna2MessageDialog1.Show("Delled successfuly..");
+                        LoadData();// Refresh data after delete
+                    }
                 }
+
             }
         }
     }
