+++ View/frmUserView.cs
@@ -21,11 +21,6 @@ namespace IM_System.View
             guna2MessageDialog1.Parent = frmMain.Instance;
         }
 
-        private void frmUserView_Load(object sender, EventArgs e)
-        {
-            LoadData();
-        }
-
         public virtual void btnAdd_Click(object sender, EventArgs e)
         {
             MainClass.BlurBackground(new frmUserAdd());
@@ -54,6 +49,7 @@ namespace IM_System.View
 
         private void guna2DataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
         {
+
             //Update
             if (guna2DataGridView1.CurrentCell.OwningColumn.Name == "dgvEdit")
             {
@@ -65,24 +61,33 @@ namespace IM_System.View
                 frm.txtPhone.Text = Convert.ToString(guna2DataGridView1.CurrentRow.Cells["dgvphon"].Value);
 
                 MainClass.BlurBackground(frm);
+                LoadData();
             }
             //Delete
             if (guna2DataGridView1.CurrentCell.OwningColumn.Name == "dgvDel")
             {
-                int id = Convert.ToInt32(guna2DataGridView1.CurrentRow.Cells["dgvid"].Value);
-                string qry = "Delete from users where userID = " + id + "";
-                Hashtable ht = new Hashtable();
-                if (MainClass.SQl(qry, ht) > 0)
+
+                //Confirm Before delete
+                guna2MessageDialog1.Buttons = Guna.UI2.WinForms.MessageDialogButtons.YesNo;
+                guna2MessageDialog1.Icon = Guna.UI2.WinForms.MessageDialogIcon.Information;
+
+                if (guna2MessageDialog1.Show("Are you sure you want to delete") == DialogResult.Yes)
                 {
-                    guna2MessageDialog1.Buttons = Guna.UI2.WinForms.MessageDialogButtons.OK;
-                    guna2MessageDialog1.Icon = Guna.UI2.WinForms.MessageDialogIcon.Information;
-                    guna2MessageDialog1.Show("Delled successfuly..");
-                    LoadData();
+                    int id = Convert.ToInt32(guna2DataGridView1.CurrentRow.Cells["dgvid"].Value);
+                    string qry = "Delete from users where userID = " + id + "";
+                    Hashtable ht = new Hashtable();
+                    if (MainClass.SQl(qry, ht) > 0)
+                    {
+                        guna2MessageDialog1.Buttons = Guna.UI2.WinForms.MessageDialogButtons.OK;
+                        guna2MessageDialog1.Icon = Guna.UI2.WinForms.MessageDialogIcon.Information;
+                        guna2MessageDialog1.Show("Delled successfuly..");
+                        LoadData();
+                    }
                 }
             }
         }
 
-        private void frmUserView_Load_1(object sender, EventArgs e)
+        private void frmUserView_Load(object sender, EventArgs e)
         {
             LoadData();
         }
